Changing float -> decimal in schema.

+++ openjudge/db/migrate/20121103064659_create_gifts.rb
@@ -2,11 +2,11 @@ class CreateGifts < ActiveRecord::Migration
   def self.up
     create_table :gifts do |t|
       t.string      :description
-      t.float       :value
+      t.decimal     :value
       t.date        :date_given
       t.string      :address_raw
-      t.float       :latitude
-      t.float       :longitude
+      t.decimal     :latitude
+      t.decimal     :longitude
       t.string      :image_uri
       t.integer     :donor_id
       t.integer     :judge_id
